requirements.txt
aiogram>=3.7.0,<4.0.0
SQLAlchemy[asyncio]>=2.0.30
python-dotenv>=1.0.1
aiosqlite>=0.20.0
psycopg[binary]>=3.2.0
